explore.py: implement smart wait
smart wait is a new waiting strategy in the sequential case.

The idea is to leverage the fact that certain long-running programs like
test-suites regularly output to stdout when they are running correctly.
If they stop outputing for a certain amount of time, this probably means
that they are hanging and we can already abort.

In such cases, smart wait will use the timeout as poll frequency: if
after one timeout the output hasn't changed, consider the program
hanged. If not, wait for another timeout. The value passed to the smart
wait option determines how many of such iterations are allowed.

By default, smart wait is disabled, meaning that its value is set to 1.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@manchester.ac.uk>