README.md: remove bit about reproduced artifact
This is confusing because in the end the badge was awarded to the
artifact.

Indeed, the ASPLOS'24 committee decided that the reproducibility
experiments we provided were sufficient to showcase the reproducibility
of the artifact as a whole, and that reproducing the full data set was
not necessary to award the reproduced badge.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@manchester.ac.uk>